pyproject.toml
[tool.poetry]
name = "nanoparticle-wear"
version = "0.1.0"
description = ""
authors = ["rodyherrera <contact@rodyherrera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.12.3"
numpy = "^2.2.5"
matplotlib = "^3.10.1"
ovito = "^3.12.2"
pandas = "^2.2.3"
scipy = "^1.15.2"
tqdm = "^4.67.1"
seaborn = "^0.13.2"
pyyaml = "^6.0.2"
jinja2 = "^3.1.6"
numba = "^0.61.2"
cupy-cuda12x = "^13.4.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"